STM32F3: Comp: Started converting to shifted enums.
The enums representing possible states of bits in the COMP CSR
register used to be defined as bits relative to the LSBit of the
corresponding setting. Now they are converted to the actual bit
pattern that is aligned to the CSR register.
This makes reading and writing them to/from the CSR register
easier since no shifting for a read/write is required anymore.