STM32: Hacked F2/F4 driver to generate one class for every gpio pin.
The new implementation is not 100% backwards compatible
because the PullType, OutputType, OutputSpeed and AlternateFunction
enums were moved to the Gpio Base class to make things more
readable (getting rid of all the ::xpcc::stm32 prefixes)

This new driver probably does not compile right now.
Nonetheless it is important as it serves as a foundation to
build the new template build system around.