Tools: New device file tool that can parse device and driver xml
The Tool then creates a build list containing input files
with their specific output files and for templates
a substitution dictionary.

As a small test one can try running device.py as a stand alone
python script.

The code is somewhat hacked together so it still needs some work
if it is supposed to be used in XPCC.