Tool: Device Files: Driver data can be exchanged using dicts
DeviceFile.getDriverList return a list of dicts that
contain all data about a driver that can be extracted
from the xml device file.
This dict can be used to create a DriverFile Object
(using DriverFile.fromDict).
From this DriverFile instance you can call "getBuildList"
to get a list of static and template files that
depend on the driver.