Tools: DeviceFiles: Added Parameters for Drivers
This makes it possible to specify certain parameters for drivers.
All you need to do is define the parameter in the driver xml file.
E.g.:
<parameter name="enum_parameter" type="enum" values="A;B;C;default">default</parameter>
<parameter name="buffer_size" type="int" min="0" max="256">default</parameter>
<parameter name="feature_enable" type="bool">true</parameter>

Now you can set default values on a device basis in the xml device file:
<driver type="test" name="name_of_the_driver_that_needs_parameter">
	<parameter name="enum_parameter">A</parameter>
	<parameter name="feature_enable">0</parameter>
</driver>

In addition to that the user will be able to override parameters in the project.cfg
file. This still needs to be implemented.