Clock: create input source classes using jinja2 macros
Follwing the first great virtue of a programmer:
The template constructs for the static clock interface
are rather complex but do not differ much on various
platforms.
So we safe some time by just specifying the things that are
actually different.
Maybe the macros will need to be adjusted a little bit when
new platforms.