Clock: extracted common glue code for the static interface
The common code now resides in clock/common/static.macros.
The are probably some STM32 specific things in there that
will need to be adjusted as we go.

This commit also renames the Clock class that was renamed Stm32Clock
some days ago to the more common ClockTree.
It needs to be discuessed if we want to rename it back to Clock.
For now ClockTree id ok in order to distingush between ClockTree
and the common XPCC clock code.