SconsTools: add some module support
This is not the module support that
we wanted, but it's the module support
that we can get with one hour of work.
Currently you can place a module.xml
file in the top directory of your module
and give it a name, like this:

<?xml version='1.0' encoding='UTF-8' ?>
<!DOCTYPE rca SYSTEM "../../xml/module.dtd">
<rca version="1.0">
	<module name="freertos" />
</rca>

All modules will be excluded from being built
on all plattforms by default.
To include them, add a modules section to
your project config.
project.cfg:
[modules]
freertos = include