xpcc communication: parser: support additional include search paths
In addition to searching in the root xml file's directory for
includes, the -I flag can be used to add other directories to
the include path.

Please not: this currently is not integreated into the
system_design scons tool, thus you will not be able
to use this feature with scons at the moment.
However, if you wanted to you could call the
builder files manually.