Extend xpcc::Clock, Timestamp, Timeout and PeriodicTimer to 32 bits.
For backwards compatibility, the 32bit classes are prefixed with `Long`.
So the new classes are `LongTimestamp`, `LongTimeout` and 
`LongPeriodicTimer`.
Existing code does not have to be modified and behaves as before.

xpcc::Clock::now() is a template function, which is specialized for
`Timestamp` and `LongTimestamp` and defaults to the 16bit `Timestamp`.
If you need to manually access the 32bit clock value, you must 
explicitly specify the `LongTimestamp`:
`xpcc::LongTimestamp stamp = xpcc::Clock::now<xpcc::LongTimestamp>();`

See issue #17 for a detailed discussion.