Architecture: Core/Clock: Add dynamic delay function for nanoseconds.
Tested on the STM32F3 with data and function in instruction CCM.
Minimum delay is about 375ns @ 72MHz.
The functions adapts its minimum delay for current clock rate,
deprecating reliance on the F_CPU macro and enabling clock switching
at runtime.

However, unlike `delay_us` and `delay_ms`, which measure the elapsed
time using the DWTs cycle counter, this methods executes a fixed number
of cycles, and its timing is therefore not accurate when interrupted
by an exception.

Unfortunately, the setup cost varies anywhere from 19 cycles to 37
cycles, so this is not the most accurate method to start with, however,
further optimization would allow the entire `_delay_ns` function to be
inline as it only has 16 bytes total size, and this could potentially
amoratize setup costs.