Architecture: Linker: Place .fastcode into ROM. Generate seperate .vector section.
The problem is that code in SRAM is accessed through the S-Bus which
incurs one wait-stage.
It is therefore faster to execute code from Flash which is accessed
through the I-Bus via the ART Accelerator.

Only for devices which can execute code from the CCM (like the STM32F3)
the .fastcode section is placed into CCM RAM.