Architecture: Linker: Make heap and stack grow away from each other. StartUp Script: Use proper stack pointers for MSP and PSP!
We use the main stack in both handler and thread mode!
Therefore this renames `__handler/__thread` to `__main/__process`.

If the heap or stack overflows a hard-fault is now generated instead of
just silently corrupting each other! \o/
The hard-fault handler switches to the process stack pointer so that the
entire main stack is left unmodified and therefore preserves most stack
information for the debugger.

The ARMv7-M has alignment restrictions for relocating the vector table!
All STM32 with CCM move either the vector table or the stacks into CCM,
so the vector table is always aligned.
For STM32 without CCM, the main stack size is increased until the vector
table offset address is aligned properly.
Since all STM32 have less than 128 vectors, the alignment is 1kB.
Keep this in mind when modifying the main or process stack size.