Architecture: Cortex: Rewrite HardFault handler to preserve registers.
The hard fault handler switches to the process stack and pushes all
registers (r0-r12) onto this stack.
It also reads the fault registers from the SCB and preserves them too.

Then the LR, PC and PSR are opportunistically copied from the main stack
onto the process stack, however, the values are only valid, if no stack
overflow occurred before or during stacking of the hard fault exception.

The reporting code of the handler now dumps all stacked registers and 
the last part of the main stack.