Architecture: Cortex: Add stack execution trap.
The startup code constructs a trampoline in the last 12B of the stack,
which loads and branches to the hard fault handler.
The rest of the stack is initialized with relative branching 
instructions equal to `b __stack_start`.
Since only relative branches are used, the physical address of the 
stack is irrelevant.

Extra care is takes to respect the max. negative range of the immediate
operand in the `b` instruction.
If the CPU attempts to execute unused stack, it will branch to the
bottom of the stack to the trampoline either in one instruction, or if
the stack is very large and the executed address is larger than the
max. negative range, it will take 2 or more branches to reach the
hard fault handler trampoline.

The current implementation branches directly to the hard fault handler.
Ideally it would branch to a shim function which would set a global flag
to indicate a stack execution fault and then generate a hard fault
interrupt in software using the Software Trigger Interrupt Register.