Travis CI: Enable containerized build environment.
Containers spin up faster (in a few seconds) and provide more 
computational resources than the legacy build environment
Combined with the ability to cache the downloaded toolchains much faster
setup times are achieved.

Limitations of the containers are the lack of `sudo` and not being able
to add repositories to apt-get.
This is circumvented by downloading precompiled toolchains and by
building boost >v1.49 from source and installing it locally.

Download and build are cached, so this does not to happen every time.