[driver] CAN: move bit timing calculations to separate file
This is similar to how we do baudrate calculation for
UART in the stm32 driver.

As CAN seems to be very standardized, with bs1 and bs2 values
defined in the standard, the CAN calculations are placed in
generic.

This commit also adds unittests for all precalculated values,
thus we can move to calculating these values on the fly
without the risk of introducing new bugs.