[openocd] Manually reset `C_DEBUGEN` after reset.
The hard fault handler triggers a breakpoint when `C_DEBUGEN` is set,
which signals that a debugger is connected.
When no debugger is connected, the handler will continue to failure
logging or LED blinking.

OpenOcd does not reset the `C_DEBUGEN` bit on the `shutdown` command.
A workaround is to issue `reset halt`, then manually reset the bits
using `mww 0xE000EDF0 0xA05F0000`. This will also clear the `C_HALT`
bit and therefore this is equivalent to `reset run`.