[cortex] Use heap table for initialization.
For the TLSF implementation, the entire table is
used to find the best matching memory with the
requested trait.

For the newlib and block allocator implementation
the heap table is used to find the largest heap
at boot time.
This heap is then used to service the sbrk
interface and to initialize the block allocator.