[architecture] Properly namespace util macros.
To avoid name clashes with external code, the util macros are prefixed
with `XPCC_`. Compiler attribute macros are prefixed with `xpcc_` and
made lowercase. Proper documentation is added.

For backwards compatibility, the old macros are aliased to the new ones.
This can be disabled by defining `XPCC_DISABLE_LEGACY_MACROS`.