[XPCC] Require Identifier for containers.
In the progress of adding raw Ethernet frames as a backend for XPCC
communication, each container requires a unique id.

When using CAN as a backend filtering is done by component id.
Normally, each CAN controller has enough filters to accommodate up to
ten components per container. E.g. STM32F4 has 14 CAN filters.

When moving to Ethernet, the Ethernet MAC controller only has two or
six filters which may be not enough. The filters are more difficult to
configure.

The fifth byte of the MAC address will be used for the container Id and
the sixth byte for the component Id.
MAC filtering then can be easily implemented by filtering for the first
five bytes of the MAC address.

There is no overhead for CAN communication as container Id is omitted.