[XPCC] Fix ZeroMQ backend
This change fixes the following problems in the ZeroMQ backend:

The previous implementation receives only one packet for each call to
update(). This severely limits the received packet throughput in all code that
calls update() in fixed intervals.

Receiving a too short (invalid) packet may cause dereferencing of
already deleted memory since the packet is accepted before a length
check is performed. Also the implementation does not check for the maximum
payload size that can be stored in an xpcc::SmartPointer.

The dropPacket() function does not check for the availability of a packet.
Thus calling dropPacket() twice will cause a double delete and the program may
crash.

An xpcc::SmartPointer with a data size equal to the total size of the
received ZeroMQ message is allocated to store the message payload. As this size
also contains the header size, each allocated buffer is larger than necessary.