[XPCC] ZeroMQ: Check maximum size of payloads to be sent
The current implementation does not check for the maximum size of
packets to be sent. The packet payload is an xpcc::SmartPointer whose
size may not exceed 65529 bytes (according to documentation). Due to
the absence of size checks in xpcc::SmartPointer it is possible to
create a SmartPointer up to the size of 2^16-1 = 65535 which will
overflow an uint16_t addition in ZeroMQConnector::sendPacket().