[architecture] Make bool conversion operator of Flags explicit
The bool conversion of xpcc::Register and thus all derived flags classes
is implemented in terms of "explicit operator bool()". This fixes issues
with function overloading on flags parameters.

Implicit bool conversion without an explicit cast is only applied in
the following contexts:
* conditions of if, while, for, do-while statements
* logical operators (&&, ||)
* negation (operator !)
* static_assert

This will probably break user code that relies on implicit bool
conversion of flags in return statements or in assignments to bool
variables.