[clock] Exchange for mock-up during testing
modm::Clock is used throughout the code base for a milli-second tick and
its implementation is platform-dependent and therefore needs to be
replaced completely for testing.

The implementation of modm::Clock::increment() is moved from interface
to platform so that during testing on AVR or ARM Cortex-M, the testing
code can choose to ignore calls to it, e.g. from SysTick interrupt.