[build.scons] Improve and add git info tool
The git tool generated a header file <xpcc_git_info.hpp> which was
explicitly #included. However, in modm, all generated files are to be
checked into the user git repository, which would not be useful for
this tools, since it would overwrite the file with (almost) every build
and therefore change repository state.

The CPP defines are now passed to the compiler as command line arguments
to avoid this problem.