[gpio] Remove separate Gpio{Out,In}put types
By having distinct Output and Input types for the GPIO classes it was
impossible to accidentally misconfigure an output class as input and vice
versa. However in practice these "protections" weren't very strong since there
was no global way to assert that a pin was only used as an output or input.
In addition the implementation complexity of having multiple types describe
the same hardware is just not worth it.