[I2S] Add audio DAC driver and interrupt driven I2S
I2S master driver can be used with the TXE (transmit buffer empty)
to transmit audio samples.

The DAC driver is not that flexible at this moment. It just provides
enough config that it can output audio on the headphone output.
Register field definition is also not complete. I just added that
much that is currently
enough to make it work.