[samg] Add faster transmit only method to SPI driver
Synchronous transfer is slowed down by the having to wait for a complete
transfer to read the received byte. When only transmit is required, this
method allows for writing the next transmit byte while the current byte
is being written out to increase throughput.

Also, removes unnecessary setOutput calls in connect method.