[stm32] [dma] [irq] Check for pointer first for speed
In case of IRQ for transferComplete is used and halfTransferComplete is
not used:

HT_Flag is always/mostly set when TC_Flag is set.
So checking the flag first is wasting time.

In case that TC_Flag is set and halfTransferComplete is used, IRQ
latency stays the same with this change.