[stm32] Place .data section into D1_SRAM instead of DTCM for H7
The data sections were previously placed into the first memory of the
contiguous section with the lowest address. The selection is changed
such that the first memory of the largest contiguous ram section is
used. For all supported Cortex-M devices except STM32H7 the resulting
linker script will be identical.

On STM32H7 .data is currently placed into the DTCM which is not
accessible by peripheral DMA transfers. With the new method the D1 AXI
SRAM is used for .data while .fastdata remains in the DTCM.